Crate state

metw-accounts-center state

This crate implements the persistence and integration interfaces defined by the service crate.

Setup Recommendations

The token consumption and email rate limiting is enforced by Redis, in case of Redis state loss, one-time tokens can be accepted again. Make sure you have enabled persistent storage in Redis.

Development Notes

Concurrent transactions are not supported on multiplexed Redis connections. Use a dedicated connection for each operation that requires a Redis transaction.

Modules

redis_keystestutil

Redis keys used with repositories.

Structs

AccountRepoImpl

Account repository using PostgreSQL.

ApplicationRepoImpl

Application repository using PostgreSQL.

AuthorizationCodeRepoImpl

Authorization code repository using Redis.

CaptchaClientImpl

CAPTCHA client validates Cloudflare’s Turnstile.

Config

Config holds the configuration for the application.

EmailClientImpl

Email client for sending emails.

EmailLimitingRepoImpl

Email limiting repository using Redis.

TokenRepoImpl

Token repository using Redis.