Crate state

metw-accounts-center state

This crate contains the state, side effects, of the service crate. External integrations and data repository is served by this crate.

Setup Recommendations

The token consumption and email rate limiting is enforced by Redis, in case of Redis state loss, one-time tokens can be accepted again. Make sure you have enabled persistent storage in Redis.

Modules

redis_keystestutil

Redis keys used with repositories.

Structs

AccountRepoImpl

Account repository using PostgreSQL.

AwsEmailClientImplemail-client-aws

Email client for sending emails.

CaptchaClientImpl

CAPTCHA client validates Cloudflare’s Turnstile.

Config

Config holds the configuration for the application.

EmailLimitingRepoImpl

Email limiting repository using Redis.

LettreEmailClientImplemail-client-lettre

Email client for sending emails.

TokenRepoImpl

Token repository using Redis.